fix build, move some lil extensions back
